pyproject.toml
[tool.poetry]
name = "reddit-saved-pptx"
version = "0.1.0"
description = ""
authors = ["strNophix <nvdpoel01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
praw = "^7.6.0"
Jinja2 = "^3.1.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
